Move all mod constants to its own folder
