--- StardewArchipelago/Constants/ArchipelagoSkillNames.cs
@@ -1,12 +1,0 @@
-﻿namespace StardewArchipelago.Constants
-{
-    public class ArchipelagoSkillNames
-    {
-        public const string LUCK = "Luck";
-        public const string BINNING = "Binning";
-        public const string MAGIC = "Magic";
-        public const string ARCHAEOLOGY = "Archaeology";
-        public const string COOKING = "Cooking";
-        public const string SOCIALIZING = "Socializing";
-    }
-}
