+++ StardewArchipelago/Constants/Modded/ArchipelagoSkillNames.cs
@@ -1,0 +1,12 @@
+﻿namespace StardewArchipelago.Constants.Modded
+{
+    public class ArchipelagoSkillNames
+    {
+        public const string LUCK = "Luck";
+        public const string BINNING = "Binning";
+        public const string MAGIC = "Magic";
+        public const string ARCHAEOLOGY = "Archaeology";
+        public const string COOKING = "Cooking";
+        public const string SOCIALIZING = "Socializing";
+    }
+}
