+++ StardewArchipelago/Constants/Modded/ModInternalNames.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+
+namespace StardewArchipelago.Constants.Modded
+{
+    public static class ModInternalNames
+    {
+        public static readonly Dictionary<string, string> InternalNames = new Dictionary<string, string>()
+        {
+            { ModNames.ALEC, "NPC Alec Revisited" },
+            { ModNames.ARCHAEOLOGY, "ExcavationSkill" },
+            { ModNames.AYEISHA, "Ayeisha" },
+            { ModNames.DELORES, "Delores" },
+            { ModNames.EUGENE, "Eugene NPC Eng Translation" },
+            { ModNames.JASPER, "NPC Jasper" },
+            { ModNames.JUNA, "NPC Juna" },
+            { ModNames.MISTER_GINGER, "NPC Mr Ginger" },
+            { ModNames.RILEY, "NPC Riley" },
+            { ModNames.SHIKO, "Papaya's Custom NPC Mod" },
+            { ModNames.WELLWICK, "[CP] Wellwick" },
+            { ModNames.YOBA, "[CP] Yoba NPC" },
+        };
+    }
+}
