+++ StardewArchipelago/Constants/Modded/ModNames.cs
@@ -1,0 +1,35 @@
+﻿namespace StardewArchipelago.Constants.Modded
+{
+    public class ModNames
+    {
+        public const string ALEC = "Alec Revisited";
+        public const string ARCHAEOLOGY = "Archaeology";
+        public const string AYEISHA = "Ayeisha - The Postal Worker (Custom NPC)";
+        public const string BIGGER_BACKPACK = "Bigger Backpack";
+        public const string BINNING = "Binning Skill";
+        public const string COOKING = "Cooking Skill";
+        public const string DEEP_WOODS = "DeepWoods";
+        public const string DELORES = "Delores - Custom NPC";
+        public const string EUGENE = "Custom NPC Eugene";
+        public const string JASPER = "Professor Jasper Thomas";
+        public const string JUNA = "Juna - Roommate NPC";
+        public const string LUCK = "Luck Skill";
+        public const string MAGIC = "Magic";
+        public const string MISTER_GINGER = "Mister Ginger (cat npc)";
+        public const string RILEY = "Custom NPC - Riley";
+        public const string SHIKO = "Shiko - New Custom NPC";
+        public const string SKULL_CAVERN_ELEVATOR = "Skull Cavern Elevator";
+        public const string SOCIALIZING = "Socializing Skill";
+        public const string TRACTOR = "Tractor Mod";
+        public const string WELLWICK = "'Prophet' Wellwick";
+        public const string YOBA = "Custom NPC - Yoba";
+        public const string SVE = "Stardew Valley Expanded";
+        public const string ALECTO = "Alecto the Witch";
+        public const string DISTANT_LANDS = "Distant Lands - Witch Swamp Overhaul";
+        public const string LACEY = "Hat Mouse Lacey";
+        public const string BOARDING_HOUSE = "Boarding House and Bus Stop Extension";
+
+        public const string AP_SVE = "StardewAP - SVE Patch";
+        public const string AP_JASPER = "StardewAP - Jasper Patch";
+    }
+}
