+++ StardewArchipelago/GameModifications/Buildings/CarpenterMenuArchipelago.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.Locations.CodeInjections.Vanilla;
 using StardewModdingAPI;
 using StardewValley;
