+++ StardewArchipelago/GameModifications/CodeInjections/Television/LivinOffTheLandInjections.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Archipelago.MultiClient.Net.Models;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Objects;
