+++ StardewArchipelago/GameModifications/Modded/InitialModGameStateInitializer.cs
@@ -1,6 +1,6 @@
 using StardewModdingAPI;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewValley;
 
 namespace StardewArchipelago.Locations.CodeInjections.Modded
