+++ StardewArchipelago/GameModifications/Modded/ModRandomizedLogicPatcher.cs
@@ -1,7 +1,7 @@
 
 using HarmonyLib;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.GameModifications.CodeInjections.Modded;
 using StardewArchipelago.Stardew;
 using StardewModdingAPI;
