+++ StardewArchipelago/Items/ItemManager.cs
@@ -20,10 +20,10 @@ namespace StardewArchipelago.Items
         private Mailman _mail;
         private HashSet<ReceivedItem> _itemsAlreadyProcessed;
 
-        public ItemManager(IMonitor monitor, IModHelper helper, Harmony harmony, ArchipelagoClient archipelago, StardewItemManager itemManager, Mailman mail, TileChooser tileChooser, BabyBirther babyBirther, GiftSender giftSender, IEnumerable<ReceivedItem> itemsAlreadyProcessed, ModdedListsAndDictionaries moddedListsAndDictionaries)
+        public ItemManager(IMonitor monitor, IModHelper helper, Harmony harmony, ArchipelagoClient archipelago, StardewItemManager itemManager, Mailman mail, TileChooser tileChooser, BabyBirther babyBirther, GiftSender giftSender, IEnumerable<ReceivedItem> itemsAlreadyProcessed)
         {
             _archipelago = archipelago;
-            _itemParser = new ItemParser(monitor, helper, harmony, archipelago, itemManager, tileChooser, babyBirther, giftSender, moddedListsAndDictionaries);
+            _itemParser = new ItemParser(monitor, helper, harmony, archipelago, itemManager, tileChooser, babyBirther, giftSender);
             _mail = mail;
             _itemsAlreadyProcessed = itemsAlreadyProcessed.ToHashSet();
         }
