+++ StardewArchipelago/Items/ItemParser.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using HarmonyLib;
@@ -7,7 +7,7 @@ using StardewArchipelago.Archipelago.Gifting;
 using StardewArchipelago.Items.Mail;
 using StardewArchipelago.Items.Traps;
 using StardewArchipelago.Items.Unlocks;
-using StardewArchipelago.Locations.CodeInjections.Modded;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.Stardew;
 using StardewArchipelago.Stardew.NameMapping;
 using StardewModdingAPI;
@@ -26,15 +26,13 @@ namespace StardewArchipelago.Items
 
         // When More mods start to need name mapping, we can make a generic version of this
         private CompoundNameMapper _nameMapper;
-        private Dictionary<string,string> _archaeologyNameToDisplayName;
         
-        public ItemParser(IMonitor monitor, IModHelper helper, Harmony harmony, ArchipelagoClient archipelago, StardewItemManager itemManager, TileChooser tileChooser, BabyBirther babyBirther, GiftSender giftSender, ModdedListsAndDictionaries moddedListsAndDictionaries)
+        public ItemParser(IMonitor monitor, IModHelper helper, Harmony harmony, ArchipelagoClient archipelago, StardewItemManager itemManager, TileChooser tileChooser, BabyBirther babyBirther, GiftSender giftSender)
         {
             _itemManager = itemManager;
             _unlockManager = new UnlockManager(archipelago);
             _trapManager = new TrapManager(monitor, helper, harmony, archipelago, tileChooser, babyBirther, giftSender);
             _nameMapper = new CompoundNameMapper(archipelago.SlotData);
-            _archaeologyNameToDisplayName = moddedListsAndDictionaries.ArchaeologyAPNametoActualName;
         }
 
         public TrapManager TrapManager => _trapManager;
