+++ StardewArchipelago/Items/Unlocks/UnlockManager.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using StardewArchipelago.Archipelago;
 using StardewArchipelago.Items.Mail;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 
 namespace StardewArchipelago.Items.Unlocks
 {
