+++ StardewArchipelago/Locations/CodeInjections/Vanilla/MonsterSlayer/WeaponsManager.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.Stardew;
 using StardewValley.Objects;
 using StardewValley.Tools;
