+++ StardewArchipelago/Locations/CodeInjections/Vanilla/RecipePurchaseInjections.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Archipelago.MultiClient.Net.Models;
 using Microsoft.Xna.Framework;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.Stardew;
 using StardewModdingAPI;
 using StardewValley;
