+++ StardewArchipelago/Locations/CodeInjections/Vanilla/SkillInjections.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewModdingAPI;
 using StardewValley;
