+++ StardewArchipelago/Locations/Patcher/ModLocationPatcher.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System;
 using StardewArchipelago.Archipelago;
-using StardewArchipelago.Constants;
+using StardewArchipelago.Constants.Modded;
 using StardewArchipelago.GameModifications.CodeInjections.Modded;
 using StardewArchipelago.Locations.CodeInjections.Modded;
 using StardewArchipelago.Locations.CodeInjections.Modded.SVE;
